Migrate to ES6 modules
