+++ src/constants.js
@@ -1,0 +1,2 @@
+export const year = 2022;
+export const shortYear = year % 100;
